requirements.txt
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74